+++ src/array/index.ts
@@ -1,3 +1,4 @@
 export * from './array.js'
 export * from './items.js'
 export * from './maxItems.js'
+export * from './minItems.js'